requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==1.11
django-cors-headers==1.1.0
django-filter==2.0.0
djangorestframework==3.3.3
idna==2.7
mysqlclient==1.3.13
requests==2.19.1
urllib3==1.23
wincertstore==0.2
django-debug-toolbar 
docker
Celery
redis